Add test for the CardListViewControler It start with a simple snapshot
