--- magicbyirineu/Modules/CardList/Datasources & Delegates/CategoriesCollectionViewDelegate.swift
@@ -1,14 +1,0 @@
-//
-//  CategoriesCollectionViewDelegate.swift
-//  magicbyirineu
-//
-//  Created by andre.antonio.filho on 18/02/19.
-//  Copyright © 2019 DanielLima. All rights reserved.
-//
-
-import UIKit
-
-class CategoriesCollectionViewDelegate: NSObject, UICollectionViewDelegate{
-    
-}
-
